kill all the T+D stuff from bigquery
