+++ airbyte-cdk/bulk/toolkits/load-db/src/main/kotlin/io/airbyte/cdk/load/orchestration/db/Sql.kt
@@ -111,5 +111,7 @@ data class Sql(val transactions: List<List<String>>) {
                     .filter { it.isNotEmpty() }
             )
         }
+
+        fun empty() = Sql(emptyList())
     }
 }
