+++ airbyte-integrations/connectors/destination-bigquery/src/main/kotlin/io/airbyte/integrations/destination/bigquery/write/typing_deduping/direct_load_tables/BigqueryDirectLoadSqlGenerator.kt
@@ -4,16 +4,32 @@
 
 package io.airbyte.integrations.destination.bigquery.write.typing_deduping.direct_load_tables
 
+import com.google.cloud.bigquery.StandardSQLTypeName
 import io.airbyte.cdk.load.command.Dedupe
 import io.airbyte.cdk.load.command.DestinationStream
+import io.airbyte.cdk.load.data.AirbyteType
+import io.airbyte.cdk.load.data.ArrayType
+import io.airbyte.cdk.load.data.ArrayTypeWithoutSchema
+import io.airbyte.cdk.load.data.BooleanType
+import io.airbyte.cdk.load.data.DateType
+import io.airbyte.cdk.load.data.IntegerType
+import io.airbyte.cdk.load.data.NumberType
+import io.airbyte.cdk.load.data.ObjectType
+import io.airbyte.cdk.load.data.ObjectTypeWithEmptySchema
+import io.airbyte.cdk.load.data.ObjectTypeWithoutSchema
+import io.airbyte.cdk.load.data.StringType
+import io.airbyte.cdk.load.data.TimeTypeWithTimezone
+import io.airbyte.cdk.load.data.TimeTypeWithoutTimezone
+import io.airbyte.cdk.load.data.TimestampTypeWithTimezone
+import io.airbyte.cdk.load.data.TimestampTypeWithoutTimezone
+import io.airbyte.cdk.load.data.UnionType
+import io.airbyte.cdk.load.data.UnknownType
 import io.airbyte.cdk.load.orchestration.db.CDC_DELETED_AT_COLUMN
 import io.airbyte.cdk.load.orchestration.db.ColumnNameMapping
 import io.airbyte.cdk.load.orchestration.db.Sql
 import io.airbyte.cdk.load.orchestration.db.TableName
 import io.airbyte.cdk.load.orchestration.db.direct_load_table.DirectLoadSqlGenerator
-import io.airbyte.integrations.destination.bigquery.write.typing_deduping.legacy_raw_tables.BigQueryTypingDedupingSqlGenerator.Companion.QUOTE
-import io.airbyte.integrations.destination.bigquery.write.typing_deduping.legacy_raw_tables.BigQueryTypingDedupingSqlGenerator.Companion.clusteringColumns
-import io.airbyte.integrations.destination.bigquery.write.typing_deduping.legacy_raw_tables.BigQueryTypingDedupingSqlGenerator.Companion.toDialectType
+import java.util.ArrayList
 import java.util.stream.Collectors
 import org.apache.commons.lang3.StringUtils
 
@@ -218,7 +234,7 @@ class BigqueryDirectLoadSqlGenerator(private val projectId: String?) : DirectLoa
             if (importType.cursor.isEmpty()) {
                 ""
             } else if (importType.cursor.size == 1) {
-                val columnName = importType.cursor.first()
+                val columnName = columnNameMapping[importType.cursor.first()]!!
                 "`$columnName` DESC NULLS LAST"
             } else {
                 throw UnsupportedOperationException(
@@ -245,5 +261,53 @@ class BigqueryDirectLoadSqlGenerator(private val projectId: String?) : DirectLoa
                FROM numbered_rows
                WHERE row_number = 1
                """.trimIndent()
+    }
+
+    companion object {
+        const val QUOTE: String = "`"
+
+        fun toDialectType(type: AirbyteType): StandardSQLTypeName =
+            when (type) {
+                BooleanType -> StandardSQLTypeName.BOOL
+                DateType -> StandardSQLTypeName.DATE
+                IntegerType -> StandardSQLTypeName.INT64
+                NumberType -> StandardSQLTypeName.NUMERIC
+                StringType -> StandardSQLTypeName.STRING
+                TimeTypeWithTimezone -> StandardSQLTypeName.STRING
+                TimeTypeWithoutTimezone -> StandardSQLTypeName.TIME
+                TimestampTypeWithTimezone -> StandardSQLTypeName.TIMESTAMP
+                TimestampTypeWithoutTimezone -> StandardSQLTypeName.DATETIME
+                is ArrayType,
+                ArrayTypeWithoutSchema,
+                is ObjectType,
+                ObjectTypeWithEmptySchema,
+                ObjectTypeWithoutSchema -> StandardSQLTypeName.JSON
+                is UnionType ->
+                    if (type.isLegacyUnion) {
+                        toDialectType(type.chooseType())
+                    } else {
+                        StandardSQLTypeName.JSON
+                    }
+                is UnknownType -> StandardSQLTypeName.JSON
+            }
+
+        fun clusteringColumns(
+            stream: DestinationStream,
+            columnNameMapping: ColumnNameMapping
+        ): List<String> {
+            val clusterColumns: MutableList<String> = ArrayList()
+            if (stream.importType is Dedupe) {
+                // We're doing de-duping, therefore we have a primary key.
+                // Cluster on the first 3 PK columns since BigQuery only allows up to 4 clustering
+                // columns,
+                // and we're always clustering on _airbyte_extracted_at
+                (stream.importType as Dedupe).primaryKey.stream().limit(3).forEach {
+                    pk: List<String> ->
+                    clusterColumns.add(columnNameMapping[pk.first()]!!)
+                }
+            }
+            clusterColumns.add("_airbyte_extracted_at")
+            return clusterColumns
+        }
     }
 }
