+++ airbyte-integrations/connectors/destination-bigquery/src/main/kotlin/io/airbyte/integrations/destination/bigquery/write/typing_deduping/legacy_raw_tables/BigqueryTypingDedupingDatabaseInitialStatusGatherer.kt
@@ -5,48 +5,19 @@
 package io.airbyte.integrations.destination.bigquery.write.typing_deduping.legacy_raw_tables
 
 import com.google.cloud.bigquery.BigQuery
-import com.google.cloud.bigquery.Field
 import com.google.cloud.bigquery.QueryJobConfiguration
-import com.google.cloud.bigquery.StandardSQLTypeName
-import com.google.cloud.bigquery.StandardTableDefinition
-import com.google.cloud.bigquery.TableDefinition
 import com.google.cloud.bigquery.TableId
-import com.google.cloud.bigquery.TimePartitioning
-import com.google.common.annotations.VisibleForTesting
-import io.airbyte.cdk.load.command.Append
-import io.airbyte.cdk.load.command.Dedupe
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.command.Overwrite
-import io.airbyte.cdk.load.message.Meta
-import io.airbyte.cdk.load.orchestration.db.ColumnNameMapping
 import io.airbyte.cdk.load.orchestration.db.DatabaseInitialStatusGatherer
 import io.airbyte.cdk.load.orchestration.db.TableName
 import io.airbyte.cdk.load.orchestration.db.TableNames
-import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.AlterTableReport
 import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.FinalTableInitialStatus
 import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.RawTableInitialStatus
 import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.TableCatalog
 import io.airbyte.cdk.load.orchestration.db.legacy_typing_deduping.TypingDedupingDatabaseInitialStatus
-import io.airbyte.cdk.util.CollectionUtils
-import io.github.oshai.kotlinlogging.KotlinLogging
-import java.math.BigInteger
-import java.util.stream.Collectors
-import java.util.stream.Stream
-
-private val logger = KotlinLogging.logger {}
 
 class BigqueryTypingDedupingDatabaseInitialStatusGatherer(private val bq: BigQuery) :
     DatabaseInitialStatusGatherer<TypingDedupingDatabaseInitialStatus> {
-    private fun findExistingTable(finalTableName: TableName): TableDefinition? {
-        val table = bq.getTable(finalTableName.namespace, finalTableName.name)
-        return table?.getDefinition()
-    }
-
-    private fun isFinalTableEmpty(finalTableName: TableName): Boolean {
-        return BigInteger.ZERO ==
-            bq.getTable(TableId.of(finalTableName.namespace, finalTableName.name)).numRows
-    }
-
     private fun getInitialRawTableState(
         rawTableName: TableName,
         suffix: String
@@ -113,25 +84,15 @@ class BigqueryTypingDedupingDatabaseInitialStatusGatherer(private val bq: BigQue
         streams: TableCatalog,
     ): Map<DestinationStream, TypingDedupingDatabaseInitialStatus> {
         return streams.mapValues { (stream, names) ->
-            val (tableNames, columnNameMapping) = names
-            val finalTable = findExistingTable(tableNames.finalTableName!!)
+            val (tableNames, _) = names
+            // we're never actually doing anything with the final table
+            // so just return a hardcoded "safe" status
             val finalTableStatus =
-                finalTable?.let {
-                    FinalTableInitialStatus(
-                        isSchemaMismatch =
-                            !existingSchemaMatchesStreamConfig(
-                                stream,
-                                columnNameMapping,
-                                finalTable
-                            ),
-                        isEmpty = isFinalTableEmpty(tableNames.finalTableName!!),
-                        // for now, just use 0. this means we will always use a temp final table.
-                        // platform has a workaround for this, so it's OK.
-                        // TODO only fetch this on truncate syncs
-                        // TODO once we have destination state, use that instead of a query
-                        finalTableGenerationId = 0,
-                    )
-                }
+                FinalTableInitialStatus(
+                    isSchemaMismatch = false,
+                    isEmpty = true,
+                    finalTableGenerationId = stream.generationId,
+                )
             val rawTableState = getInitialRawTableState(tableNames.rawTableName!!, "")
             val tempRawTableState =
                 getInitialRawTableState(
@@ -143,134 +104,6 @@ class BigqueryTypingDedupingDatabaseInitialStatusGatherer(private val bq: BigQue
                 rawTableState,
                 tempRawTableState,
             )
-        }
-    }
-
-    private fun existingSchemaMatchesStreamConfig(
-        stream: DestinationStream,
-        columnNameMapping: ColumnNameMapping,
-        existingTable: TableDefinition
-    ): Boolean {
-        val alterTableReport = buildAlterTableReport(stream, columnNameMapping, existingTable)
-        var tableClusteringMatches = false
-        var tablePartitioningMatches = false
-        if (existingTable is StandardTableDefinition) {
-            tableClusteringMatches = clusteringMatches(stream, columnNameMapping, existingTable)
-            tablePartitioningMatches = partitioningMatches(existingTable)
-        }
-        logger.info {
-            "Alter Table Report ${alterTableReport.columnsToAdd} ${alterTableReport.columnsToRemove} ${alterTableReport.columnsToChangeType}; Clustering $tableClusteringMatches; Partitioning $tablePartitioningMatches"
-        }
-
-        return alterTableReport.isNoOp && tableClusteringMatches && tablePartitioningMatches
-    }
-
-    internal fun buildAlterTableReport(
-        stream: DestinationStream,
-        columnNameMapping: ColumnNameMapping,
-        existingTable: TableDefinition,
-    ): AlterTableReport {
-        val pks = getPks(stream, columnNameMapping)
-
-        val streamSchema: Map<String, StandardSQLTypeName> =
-            stream.schema.asColumns().entries.associate {
-                columnNameMapping[it.key]!! to BigQueryTypingDedupingSqlGenerator.toDialectType(it.value.type)
-            }
-
-        val existingSchema =
-            existingTable.schema!!.fields.associate { it.name to it.type.standardType }
-
-        // Columns in the StreamConfig that don't exist in the TableDefinition
-        val columnsToAdd =
-            streamSchema.keys
-                .stream()
-                .filter { name: String ->
-                    !CollectionUtils.containsIgnoreCase(existingSchema.keys, name)
-                }
-                .collect(Collectors.toSet())
-
-        // Columns in the current schema that are no longer in the StreamConfig
-        val columnsToRemove =
-            existingSchema.keys
-                .stream()
-                .filter { name: String ->
-                    !CollectionUtils.containsIgnoreCase(streamSchema.keys, name) &&
-                        !CollectionUtils.containsIgnoreCase(Meta.COLUMN_NAMES, name)
-                }
-                .collect(Collectors.toSet())
-
-        // Columns that are typed differently than the StreamConfig
-        val columnsToChangeType =
-            Stream.concat(
-                    streamSchema.keys
-                        .stream() // If it's not in the existing schema, it should already be in the
-                        // columnsToAdd Set
-                        .filter { name: String ->
-                            CollectionUtils.matchingKey(
-                                    existingSchema.keys,
-                                    name
-                                ) // if it does exist, only include it in this set if the type (the
-                                // value in each respective map)
-                                // is different between the stream and existing schemas
-                                .map { key: String ->
-                                    existingSchema[key] != streamSchema[name]
-                                } // if there is no matching key, then don't include it because it
-                                // is probably already in columnsToAdd
-                                .orElse(false)
-                        }, // OR columns that used to have a non-null constraint and shouldn't
-                    // (https://github.com/airbytehq/airbyte/pull/31082)
-
-                    existingTable.schema!!
-                        .fields
-                        .stream()
-                        .filter { pks.contains(it.name) && it.mode == Field.Mode.REQUIRED }
-                        .map { obj: Field -> obj.name }
-                )
-                .collect(Collectors.toSet())
-
-        return AlterTableReport(
-            columnsToAdd,
-            columnsToRemove,
-            columnsToChangeType,
-        )
-    }
-
-    companion object {
-        @VisibleForTesting
-        fun clusteringMatches(
-            stream: DestinationStream,
-            columnNameMapping: ColumnNameMapping,
-            existingTable: StandardTableDefinition,
-        ): Boolean {
-            return (existingTable.clustering != null &&
-                CollectionUtils.containsAllIgnoreCase(
-                    HashSet<String>(existingTable.clustering!!.fields),
-                    BigQueryTypingDedupingSqlGenerator.clusteringColumns(stream, columnNameMapping)
-                ))
-        }
-
-        @VisibleForTesting
-        fun partitioningMatches(existingTable: StandardTableDefinition): Boolean {
-            return existingTable.timePartitioning != null &&
-                existingTable.timePartitioning!!
-                    .field
-                    .equals("_airbyte_extracted_at", ignoreCase = true) &&
-                TimePartitioning.Type.DAY == existingTable.timePartitioning!!.type
-        }
-
-        private fun getPks(
-            stream: DestinationStream,
-            columnNameMapping: ColumnNameMapping
-        ): Set<String> {
-            return when (stream.importType) {
-                Append,
-                Overwrite -> emptySet()
-                is Dedupe ->
-                    (stream.importType as Dedupe)
-                        .primaryKey
-                        .map { pk -> columnNameMapping[pk.first()]!! }
-                        .toSet()
-            }
         }
     }
 }
